refactor(storage): split monolithic storage.rs into modular directory
Decompose the 1220-line storage.rs into three focused files:

- mod.rs (121 lines): type definitions, Database struct, new(), call_blocking()
- migration.rs (311 lines): schema DDL, versioned migrations, migration tests
- queries.rs (800 lines): all CRUD queries (chats, messages, sessions, tool_calls, LLM usage), query tests

No logic or API changes. Pure structural reorganization.
