+++ src/storage/mod.rs
@@ -1,0 +1,121 @@
+use std::path::Path;
+use std::sync::{Arc, Mutex};
+use std::time::Duration;
+
+use rusqlite::Connection;
+
+use crate::error::StorageError;
+
+mod migration;
+mod queries;
+
+/// Thread-safe SQLite database wrapper for conversation persistence.
+pub struct Database {
+    pub(crate) conn: Mutex<Connection>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct StoredMessage {
+    pub id: String,
+    pub chat_id: i64,
+    pub sender_name: String,
+    pub content: String,
+    pub is_from_bot: bool,
+    pub timestamp: String,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct SessionSummary {
+    pub chat_id: i64,
+    pub channel: String,
+    pub surface_thread: String,
+    pub chat_title: Option<String>,
+    pub last_message_time: String,
+    pub last_message_preview: Option<String>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct ChatInfo {
+    pub chat_id: i64,
+    pub channel: String,
+    pub external_chat_id: String,
+    pub chat_type: String,
+}
+
+#[derive(Debug, Clone)]
+pub(crate) struct SessionSnapshot {
+    pub messages_json: Option<String>,
+    pub updated_at: Option<String>,
+    pub recent_messages: Vec<StoredMessage>,
+}
+
+#[derive(Debug, Clone)]
+pub(crate) struct ToolCall {
+    pub id: String,
+    pub chat_id: i64,
+    pub message_id: String,
+    pub tool_name: String,
+    pub tool_input: String,
+    pub tool_output: Option<String>,
+    pub timestamp: String,
+}
+
+pub(crate) struct LlmUsageLogEntry<'a> {
+    pub chat_id: i64,
+    pub caller_channel: &'a str,
+    pub provider: &'a str,
+    pub model: &'a str,
+    pub input_tokens: i64,
+    pub output_tokens: i64,
+    pub request_kind: &'a str,
+}
+
+pub async fn call_blocking<T, F>(db: Arc<Database>, f: F) -> Result<T, StorageError>
+where
+    T: Send + 'static,
+    F: FnOnce(&Database) -> Result<T, StorageError> + Send + 'static,
+{
+    tokio::task::spawn_blocking(move || f(db.as_ref()))
+        .await
+        .map_err(|error| StorageError::TaskJoin(error.to_string()))?
+}
+
+impl Database {
+    pub(crate) fn new(db_path: &Path) -> Result<Self, StorageError> {
+        let legacy_db = db_path
+            .parent()
+            .and_then(|runtime| runtime.parent())
+            .map(|root| root.join("data").join("egopulse.db"))
+            .unwrap_or_else(|| Path::new("data").join("egopulse.db"));
+        if legacy_db.exists() && !db_path.exists() {
+            return Err(StorageError::InitFailed(format!(
+                "legacy_db_pending_migration: found {}, but {} does not exist. \
+                 run 'mv {} {}' to migrate.",
+                legacy_db.display(),
+                db_path.display(),
+                legacy_db.display(),
+                db_path.display(),
+            )));
+        }
+
+        if let Some(parent) = db_path.parent() {
+            std::fs::create_dir_all(parent)?;
+        }
+
+        let conn = Connection::open(db_path)?;
+        conn.execute_batch("PRAGMA journal_mode=WAL;")?;
+        conn.busy_timeout(Duration::from_secs(5))?;
+
+        migration::run_migrations(&conn)?;
+
+        Ok(Self {
+            conn: Mutex::new(conn),
+        })
+    }
+
+    pub(crate) fn lock_conn(&self) -> Result<std::sync::MutexGuard<'_, Connection>, StorageError> {
+        self.conn
+            .lock()
+            .map_err(|error| StorageError::InitFailed(error.to_string()))
+    }
+}
